requirements.txt
mysql-connector-python==8.0.15
pyserial==3.4
python-dotenv==0.10.1
